fix: Исправить провреку при обновлении сущностей, разделить тесты, добавить логирование в catch блок
